refactor declarative code
